Simplify the react entry file('index.tsx)', including: 1. Move the core implementation to 'App.tsx'; 2. Lazily import 'App.tsx' to ensure language strings are    properly initialised before importing.
